+++ packages/wallet-shared/src/pairing/component/PairingStatusBox/index.tsx
@@ -43,35 +43,6 @@ export function StatusBoxModal({
 }
 
 /**
- * StatusBoxWalletEmbedded is a component that displays a status icon and a title.
- * It is used to display the status of a pairing process on wallet embedded.
- *
- * @param status - The status of the pairing process.
- * @param title - The title of the pairing process.
- * @param children - The children of the pairing process.
- */
-export function StatusBoxWalletEmbedded({
-    status,
-    title,
-    client,
-    children,
-}: PropsWithChildren<{
-    status: Status;
-    title: string;
-    client?: AnyPairingClient;
-}>) {
-    return (
-        <div className={styles.statusBoxWalletEmbeddedContainer}>
-            <div className={styles.statusBox}>
-                <InnerStatusBox status={status} title={title} />
-            </div>
-            {client && <StatusBoxRetry client={client} />}
-            {children}
-        </div>
-    );
-}
-
-/**
  * StatusBoxRefresh is a component that displays a refresh button.
  * It is used to refresh the pairing process.
  */